Simplify table-export
