--- ui-frontend/packages/catalog-ui-search/src/main/webapp/react-component/table-export/container.tsx
@@ -1,121 +1,0 @@
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser
- * General Public License as published by the Free Software Foundation, either version 3 of the
- * License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details. A copy of the GNU Lesser General Public License
- * is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
-import * as React from 'react'
-import TableExportComponent from './presentation'
-import LinearProgress from '@mui/material/LinearProgress'
-import { hot } from 'react-hot-loader'
-import {
-  ExportCountInfo,
-  DownloadInfo,
-} from '../../react-component/utils/export'
-import { StartupDataStore } from '../../js/model/Startup/startup'
-type Props = {
-  selectionInterface: () => void
-  exportFormats: Option[]
-  getWarning: (exportCountInfo: ExportCountInfo) => string
-  onDownloadClick: (downloadInfo: DownloadInfo) => void
-}
-type Option = {
-  label: string
-  value: string
-}
-type State = {
-  exportSizes: Option[]
-  exportFormat: string
-  exportSize: string
-  customExportCount: number
-}
-export default hot(module)(
-  class TableExport extends React.Component<Props, State> {
-    constructor(props: Props) {
-      super(props)
-      this.state = {
-        exportSizes: [
-          {
-            label: 'Current Page',
-            value: 'currentPage',
-          },
-          {
-            label: 'All Results',
-            value: 'all',
-          },
-          {
-            label: 'Specific Number of Results',
-            value: 'custom',
-          },
-        ],
-        exportSize: 'all',
-        exportFormat: '',
-        customExportCount: StartupDataStore.Configuration.getExportLimit(),
-      }
-    }
-    handleExportFormatChange = (value: string) => {
-      this.setState({
-        exportFormat: value,
-      })
-    }
-    handleExportSizeChange = (value: string) => {
-      this.setState({
-        exportSize: value,
-      })
-    }
-    handleCustomExportCountChange = (value: number) => {
-      this.setState({ customExportCount: value })
-    }
-    render() {
-      const { exportFormat, exportSizes, exportSize, customExportCount } =
-        this.state
-      const { exportFormats, selectionInterface, onDownloadClick, getWarning } =
-        this.props
-      return exportFormats.length === 0 ? (
-        <>
-          <LinearProgress className="w-full h-2"></LinearProgress>
-        </>
-      ) : (
-        <>
-          <>
-            {exportFormats.length > 0 ? (
-              <TableExportComponent
-                exportFormatOptions={exportFormats}
-                exportFormat={exportFormat}
-                exportSizeOptions={exportSizes}
-                exportSize={exportSize}
-                handleExportFormatChange={this.handleExportFormatChange}
-                handleExportSizeChange={this.handleExportSizeChange}
-                handleCustomExportCountChange={
-                  this.handleCustomExportCountChange
-                }
-                onDownloadClick={() =>
-                  onDownloadClick({
-                    exportFormat,
-                    exportSize,
-                    selectionInterface,
-                    customExportCount,
-                  })
-                }
-                warning={getWarning({
-                  exportSize,
-                  selectionInterface,
-                  customExportCount,
-                })}
-                customExportCount={customExportCount}
-              />
-            ) : null}
-          </>
-        </>
-      )
-    }
-  }
-)
